# orders-api/openapi.yaml
openapi: 3.0.3
info:
title: Orders API
version: 0.1.0
servers:
- url: http://localhost:3002
paths:
/health:
get:
summary: Health check
responses:
'200':
description: OK

paths:
  /orders/{id}/confirm:
    put:
      summary: Confirm order
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Confirmed }
        "404": { description: Not Found }
        "409": { description: Invalid transition }

  /orders/{id}/cancel:
    put:
      summary: Cancel order (restores stock if CREATED)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Canceled }
        "404": { description: Not Found }
        "409": { description: Invalid transition }

  /orders/{id}:
    patch:
      summary: Update order status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [CONFIRMED, CANCELED]
      responses:
        "200": { description: Updated }
        "404": { description: Not Found }
        "409": { description: Invalid transition }

  /orders:
    get:
      summary: List orders
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [CREATED, CONFIRMED, CANCELED] }
        - in: query
          name: customerId
          schema: { type: integer }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: cursor
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200": { description: OK }